requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.26.2
